+++ app/src/main/res/layout/activity_main.xml
@@ -5,8 +5,8 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.google.android.gms.maps.MapView
-        android:id="@+id/map_frame"
+    <FrameLayout
+        android:id="@+id/map_fragment_container"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
